LPS-73944 Copy code from Asset Verify to a new Asset Categories Verify Process
